requirements.txt
SQLAlchemy[asyncio]
ruff
python-dotenv
asyncpg
loguru
python-telegram-bot[job-queue]
tenacity
fastapi
uvicorn